consulting-agents/requirements.txt
# Core dependencies
anthropic>=0.18.0
pydantic>=2.0.0
pyyaml>=6.0
python-frontmatter>=1.0.0
markdown>=3.5.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Optional: For vector storage (future)
# numpy>=1.24.0
# scikit-learn>=1.3.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.5.0
